@@ -1,25 +1,24 @@
-namespace Toggl.Api.Test
+namespace Toggl.Api.Test;
+
+public class Configuration
 {
-	public class Configuration
-	{
-		/// <summary>
-		///  The API key
-		/// </summary>
-		public string ApiKey { get; set; } = string.Empty;
+	/// <summary>
+	///  The API key
+	/// </summary>
+	public string ApiKey { get; set; } = string.Empty;
 
-		/// <summary>
-		/// A sample workspace name
-		/// </summary>
-		public string SampleWorkspaceName { get; set; } = string.Empty;
+	/// <summary>
+	/// A sample workspace name
+	/// </summary>
+	public string SampleWorkspaceName { get; set; } = string.Empty;
 
-		/// <summary>
-		/// A sample client name
-		/// </summary>
-		public string SampleClientName { get; set; } = string.Empty;
+	/// <summary>
+	/// A sample client name
+	/// </summary>
+	public string SampleClientName { get; set; } = string.Empty;
 
-		/// <summary>
-		/// A sample project name
-		/// </summary>
-		public string SampleProjectName { get; set; } = string.Empty;
-	}
+	/// <summary>
+	/// A sample project name
+	/// </summary>
+	public string SampleProjectName { get; set; } = string.Empty;
 }
