@@ -1,17 +1,16 @@
 ﻿using System.Collections.Generic;
 using Newtonsoft.Json;
 
-namespace Toggl.Api.DataObjects
+namespace Toggl.Api.DataObjects;
+
+public class DetailedReport : Report
 {
-	public class DetailedReport : Report
-	{
-		[JsonProperty(PropertyName = "data")]
-		public List<ReportTimeEntry>? Data { get; set; }
+	[JsonProperty(PropertyName = "data")]
+	public List<ReportTimeEntry>? Data { get; set; }
 
-		[JsonProperty(PropertyName = "total_count")]
-		public long? TotalCount { get; set; }
+	[JsonProperty(PropertyName = "total_count")]
+	public long? TotalCount { get; set; }
 
-		[JsonProperty(PropertyName = "per_page")]
-		public long? PerPage { get; set; }
-	}
+	[JsonProperty(PropertyName = "per_page")]
+	public long? PerPage { get; set; }
 }
