@@ -1,19 +1,18 @@
 ﻿using Newtonsoft.Json;
 
-namespace Toggl.Api.DataObjects
+namespace Toggl.Api.DataObjects;
+
+public class NewBlogPost : BaseDataObject
 {
-	public class NewBlogPost : BaseDataObject
-	{
-		[JsonProperty("title")]
-		public string? Title { get; set; }
+	[JsonProperty("title")]
+	public string? Title { get; set; }
 
-		[JsonProperty("url")]
-		public string? Url { get; set; }
+	[JsonProperty("url")]
+	public string? Url { get; set; }
 
-		[JsonProperty("category")]
-		public string? Category { get; set; }
+	[JsonProperty("category")]
+	public string? Category { get; set; }
 
-		[JsonProperty("pub_date")]
-		public string? PubDate { get; set; }
-	}
+	[JsonProperty("pub_date")]
+	public string? PubDate { get; set; }
 }
