@@ -1,13 +1,12 @@
 ﻿using Newtonsoft.Json;
 
-namespace Toggl.Api.DataObjects
-{
-	public class Report : BaseDataObject
-	{
-		[JsonProperty(PropertyName = "total_grand")]
-		public long? TotalGrand { get; set; }
+namespace Toggl.Api.DataObjects;
 
-		[JsonProperty(PropertyName = "total_billable")]
-		public long? TotalBillable { get; set; }
-	}
+public class Report : BaseDataObject
+{
+	[JsonProperty(PropertyName = "total_grand")]
+	public long? TotalGrand { get; set; }
+
+	[JsonProperty(PropertyName = "total_billable")]
+	public long? TotalBillable { get; set; }
 }
