@@ -1,16 +1,15 @@
 ﻿using Newtonsoft.Json;
 
-namespace Toggl.Api.DataObjects
+namespace Toggl.Api.DataObjects;
+
+public class Tag : BaseDataObject
 {
-	public class Tag : BaseDataObject
-	{
-		[JsonProperty(PropertyName = "name")]
-		public string? Name { get; set; }
+	[JsonProperty(PropertyName = "name")]
+	public string? Name { get; set; }
 
-		[JsonProperty(PropertyName = "id")]
-		public int? Id { get; set; }
+	[JsonProperty(PropertyName = "id")]
+	public int? Id { get; set; }
 
-		[JsonProperty(PropertyName = "workspace")]
-		public Workspace? Workspace { get; set; }
-	}
+	[JsonProperty(PropertyName = "workspace")]
+	public Workspace? Workspace { get; set; }
 }
