@@ -1,26 +1,25 @@
 ﻿using Newtonsoft.Json;
 using System.Collections.Generic;
 
-namespace Toggl.Api.DataObjects
+namespace Toggl.Api.DataObjects;
+
+/// <summary>
+/// https://github.com/toggl/toggl_api_docs/blob/master/chapters/users.md#users
+/// </summary>
+public class UserExtended : User
 {
-	/// <summary>
-	/// https://github.com/toggl/toggl_api_docs/blob/master/chapters/users.md#users
-	/// </summary>
-	public class UserExtended : User
-	{
-		[JsonProperty(PropertyName = "time_entries")]
-		public List<TimeEntry>? TimeEntries { get; set; }
+	[JsonProperty(PropertyName = "time_entries")]
+	public List<TimeEntry>? TimeEntries { get; set; }
 
-		[JsonProperty(PropertyName = "projects")]
-		public List<Project>? Projects { get; set; }
+	[JsonProperty(PropertyName = "projects")]
+	public List<Project>? Projects { get; set; }
 
-		[JsonProperty(PropertyName = "tags")]
-		public List<Tag>? Tags { get; set; }
+	[JsonProperty(PropertyName = "tags")]
+	public List<Tag>? Tags { get; set; }
 
-		[JsonProperty(PropertyName = "workspaces")]
-		public List<Workspace>? Workspaces { get; set; }
+	[JsonProperty(PropertyName = "workspaces")]
+	public List<Workspace>? Workspaces { get; set; }
 
-		[JsonProperty(PropertyName = "clients")]
-		public List<Client>? Clients { get; set; }
-	}
+	[JsonProperty(PropertyName = "clients")]
+	public List<Client>? Clients { get; set; }
 }
