@@ -2,12 +2,11 @@
 using Toggl.Api.DataObjects;
 using Toggl.Api.QueryObjects;
 
-namespace Toggl.Api.Interfaces
-{
+namespace Toggl.Api.Interfaces;
+
     public interface IReportServiceAsync
     {
         IApiServiceAsync TogglSrv { get; set; }
 
         Task<DetailedReport> Detailed(DetailedReportParams requestParameters);
     }
-}
