@@ -2,15 +2,14 @@
 using System.Threading.Tasks;
 using Toggl.Api.DataObjects;
 
-namespace Toggl.Api.Interfaces
+namespace Toggl.Api.Interfaces;
+
+public interface ITagServiceAsync
 {
-	public interface ITagServiceAsync
-	{
-		/// <summary>
-		/// List tag services
-		/// https://github.com/toggl/toggl_api_docs/blob/master/chapters/tasks.md
-		/// </summary>
-		/// <returns></returns>
-		Task<List<Tag>> GetAllAsync();
-	}
+	/// <summary>
+	/// List tag services
+	/// https://github.com/toggl/toggl_api_docs/blob/master/chapters/tasks.md
+	/// </summary>
+	/// <returns></returns>
+	Task<List<Tag>> GetAllAsync();
 }
