@@ -1,10 +1,9 @@
 ﻿using Newtonsoft.Json;
 
-namespace Toggl.Api.QueryObjects
+namespace Toggl.Api.QueryObjects;
+
+public class DetailedReportParams : ReportParams
 {
-	public class DetailedReportParams : ReportParams
-	{
-		[JsonProperty(PropertyName = "page")]
-		public long? Page { get; set; }
-	}
+	[JsonProperty(PropertyName = "page")]
+	public long? Page { get; set; }
 }
