@@ -2,32 +2,41 @@
 using System.Collections.Generic;
 using Toggl.Api.DataObjects;
 
-namespace Toggl.Api.QueryObjects
+namespace Toggl.Api.QueryObjects;
+
+public class TimeEntryParams : TimeEntry
 {
-	public class TimeEntryParams : TimeEntry
+	public DateTime? StartDate { get; set; }
+
+	public DateTime? EndDate { get; set; }
+
+	public TimeEntryParams()
 	{
-		public DateTime? StartDate { get; set; }
-		public DateTime? EndDate { get; set; }
-
-		public TimeEntryParams()
-		{
-			TagNames = new List<string>();
-		}
-
-		public List<KeyValuePair<string, string>> GetParameters()
-		{
-			var lst = new List<KeyValuePair<string, string>>();
-			if (StartDate.HasValue)
-				lst.Add(new KeyValuePair<string, string>("start_date", GetStartParameter()));
-			if (EndDate.HasValue)
-				lst.Add(new KeyValuePair<string, string>("end_date", GetEndIsoDate()));
-			return lst;
-		}
-
-		private string GetStartParameter() => GetIsoDate(StartDate);
-
-		private string GetEndIsoDate() => GetIsoDate(EndDate);
-
-		private string GetIsoDate(DateTime? dt) => dt.GetValueOrDefault().ToString("yyyy-MM-ddTHH:mm:sszzz");
+		TagNames = new List<string>();
 	}
+
+	public List<KeyValuePair<string, string>> GetParameters()
+	{
+		var lst = new List<KeyValuePair<string, string>>();
+		if (StartDate.HasValue)
+		{
+			lst.Add(new KeyValuePair<string, string>("start_date", GetStartParameter()));
+		}
+
+		if (EndDate.HasValue)
+		{
+			lst.Add(new KeyValuePair<string, string>("end_date", GetEndIsoDate()));
+		}
+
+		return lst;
+	}
+
+	private string GetStartParameter()
+		=> GetIsoDate(StartDate);
+
+	private string GetEndIsoDate()
+		=> GetIsoDate(EndDate);
+
+	private static string GetIsoDate(DateTime? dt)
+		=> dt.GetValueOrDefault().ToString("yyyy-MM-ddTHH:mm:sszzz");
 }
